+++ internal/ejbca/ca_pem_data_source.go
@@ -53,7 +53,6 @@ func (d *CaPemDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, 
 }
 
 func (d *CaPemDataSource) Configure(ctx context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
-	// Prevent panic if the ejbca has not been configured.
 	if req.ProviderData == nil {
 		return
 	}
@@ -71,9 +70,13 @@ func (d *CaPemDataSource) Configure(ctx context.Context, req datasource.Configur
 }
 
 func (d *CaPemDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-	var state CaPemDataSourceModel
+	if d.client == nil {
+		resp.Diagnostics.AddError("Unconfigured EJBCA client", "The EJBCA client is not configured. Please report this issue to the ejbca developers.")
+		return
+	}
 
 	// Read Terraform configuration data into the model
+	var state CaPemDataSourceModel
 	resp.Diagnostics.Append(req.Config.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {
 		return