+++ internal/ejbca/end_entity_profile_data_source_test.go
@@ -1,7 +1,6 @@
 package ejbca
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
@@ -9,11 +8,10 @@ import (
 )
 
 func TestAccEndEntityProfileDataSource(t *testing.T) {
-
-	eep, err := getRandomAuthorizedEEP()
-	if err != nil {
-		t.Fatal(err)
-	}
+	config := getAccTestConfig(t)
+    if !config.isEnterprise {
+        t.Skip("Skipping End Entity Profile Data Source Test since connected instance was not flagged as Enterprise")
+    }
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -21,9 +19,9 @@ func TestAccEndEntityProfileDataSource(t *testing.T) {
 		Steps: []resource.TestStep{
 			// Read testing
 			{
-				Config: testAccEEPDataSourceConfig(eep),
+				Config: testAccEEPDataSourceConfig(config.endEntityProfileName),
 				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("data.ejbca_end_entity_profile.test", "id", eep),
+					resource.TestCheckResourceAttr("data.ejbca_end_entity_profile.test", "id", config.endEntityProfileName),
 					resource.TestCheckResourceAttrSet("data.ejbca_end_entity_profile.test", "subject_distinguished_name_fields.#"),
 					resource.TestCheckResourceAttrSet("data.ejbca_end_entity_profile.test", "subject_alternative_name_fields.#"),
 					resource.TestCheckResourceAttrSet("data.ejbca_end_entity_profile.test", "available_certificate_profiles.#"),
@@ -36,26 +34,12 @@ func TestAccEndEntityProfileDataSource(t *testing.T) {
 
 func testAccEEPDataSourceConfig(name string) string {
 	return fmt.Sprintf(`
-    data "ejbca_end_entity_profile" "test" {
-        end_entity_profile_name = "%s"
-    }
-    `, name)
+provider "ejbca" {
+    cert_auth {}
 }
 
-func getRandomAuthorizedEEP() (string, error) {
-	client, err := createEjbcaClient()
-	if err != nil {
-		return "", err
-	}
-
-	authorizedEeps, _, err := client.V2EndentityApi.GetAuthorizedEndEntityProfiles(context.Background()).Execute()
-	if err != nil {
-		return "", err
-	}
-
-	for _, profile := range authorizedEeps.GetEndEntitieProfiles() {
-		return profile.GetName(), nil
-	}
-
-	return "", fmt.Errorf("no authorized EEP found")
+data "ejbca_end_entity_profile" "test" {
+    end_entity_profile_name = "%s"
+}
+    `, name)
 }